Add files via upload
We gather 28 household objects (such as badminton bucket, pop cans, and cloth doll) by employing UR3 robot equipped with BarrettHand that includes tactile sensors. The Multi_finger.mat combines the tactile data of three fingers in columns, which forms a 72 dimensional tactile sequence, and the 1th column represents the object labels(28 object classes). The Finger1.mat is the tactile dataset of 1th finger. The Finger2.mat is the tactile dataset of 2th finger. The Finger3.mat is the tactile dataset of 3th finger. The 1th column of them represents the object labels(28 object classes).